sms-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
playwright==1.40.0
python-dotenv==1.0.0
pydantic==2.5.0
